requirements.txt
anyio==4.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cvlib==0.2.7
cycler==0.12.1
fonttools==4.45.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageio==2.33.0
imutils==0.5.4
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
progressbar==2.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-telegram-bot==20.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==2.1.0